vsr: extra test for checksum
vsr.checksum is pretty important, as it cements our on-disk
representation and over-the-wire protocol (so, we won't be able to
easily change it).

This PR adds a "change detector" test to ensure we don't accidentally
change the result. The idea of the test is basically to hash a bunch of
random stuff, and then assert that the final hash has a specific value.

While we are at it, extend our fuzz test to do more sanity checking:

- allow arbitrary bitflips (rather than just +%= 1)
- also test avalanching (that every bit of input affects every bit of
  the output)